final/requirements.txt
pytorch
pandas
torchsummary
torchvision
